chore: 移除废弃文件并修复安全与性能问题
- 移除了旧版的 V1 API 路由 (discover, daily-report, score 等) 和相关工具类 (content-extractor, domain-evaluator)
- 修复了 Cron 任务路由中基于 Host 头鉴权可能导致的绕过漏洞，统一修改为通过 `process.env.NODE_ENV` 和 `CRON_SECRET` 验证
- 修正了 `lib/db.ts` 中索引的创建语句，使其匹配强制转换类型 `((score->>'total_score')::float)`，防止全表扫描
- 减小了向大模型传递的推文与网页文本截断长度 (从 30000 降至 15000)，降低上下文超限风险
- 将默认大模型配置升级为 `moonshot-v1-32k` 以提升对长网页的处理能力

+++ lib/db.ts
@@ -39,7 +39,7 @@ export async function initDB(): Promise<{ success: boolean; message: string }> {
 
       // 创建索引
       await sql`CREATE INDEX idx_projects_status ON projects(status);`;
-      await sql`CREATE INDEX idx_projects_score ON projects((score->>'total_score'));`;
+      await sql`CREATE INDEX idx_projects_score ON projects(((score->>'total_score')::float));`;
       await sql`CREATE INDEX idx_projects_discovered_at ON projects(discovered_at);`;
     }
